+++ src/pages/address-book/index.vue
@@ -1,0 +1,35 @@
+<template>
+  <view class="flex items-center justify-center">
+    <image
+      class="logo w-[200rpx] h-[200rpx] border border-solid border-red-400"
+      src="/static/logo.png"
+    />
+    <view class="text-area">
+      <text class="title">{{ title }}</text>
+    </view>
+  </view>
+</template>
+
+<script setup lang="ts">
+import { ref } from 'vue'
+const title = ref('Hello')
+</script>
+
+<style>
+.logo {
+  margin-top: 200rpx;
+  margin-left: auto;
+  margin-right: auto;
+  margin-bottom: 50rpx;
+}
+
+.text-area {
+  display: flex;
+  justify-content: center;
+}
+
+.title {
+  font-size: 36rpx;
+  color: #8f8f94;
+}
+</style>