+++ src/pages/mine/index.vue
@@ -1,0 +1,14 @@
+<template>
+  <view class="content">
+    <view class="text-area">
+      <text class="title">{{ title }}</text>
+    </view>
+  </view>
+</template>
+
+<script setup lang="ts">
+import { ref } from 'vue'
+const title = ref('Hello')
+</script>
+
+<style scoped lang="scss"></style>