+++ vite.config.ts
@@ -1,7 +1,20 @@
 import { defineConfig } from 'vite'
 import uni from '@dcloudio/vite-plugin-uni'
+import tailwindcss from 'tailwindcss'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [uni()],
+  css: {
+    postcss: {
+      plugins: [tailwindcss],
+    },
+    preprocessorOptions: {
+      scss: {
+        additionalData: `
+          @use "./src/styles/var.scss" as *;
+        `,
+      },
+    },
+  },
 })