+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2024年12月6日
 - fix: 修复uniapp不支持 ES Module 形式构建问题
+- feat: 添加 tailwindcss 和 tabbar 配置
 
 ## 2024年12月5日
 - init: 初始化项目